Cliente/requirements.txt
Flask==2.2.2
requests==2.28.1
Flask-SQLAlchemy==3.0.2
SQLAlchemy==1.4.46
bcrypt==4.0.1
grpcio==1.48.2
protobuf==4.21.6
pytest==7.1.3
PyJWT==2.4.0
cryptography==37.0.2
pybreaker==1.0.0